Don't update retry state when failed to get ack response
Ack failure is not buffer flush error so it should not update retry state.
The problematic node is disabled and its node is excluded at next flush after rollback.

To resolve this problem, we add update_retry parameter to Output#rollback_write.
This parameter controls rollback behaviour considers retry or not.